account_svc_grpc/requirements.txt
flask
flask_sqlalchemy
pika
reportlab
grpcio 
grpcio-tools
requests